==> 00_scripts_raw_data/00_data_preprocessing/05_balance_dataset_oversampling-SMOTE.ipynb <==
# https://www.sharpsightlabs.com/blog/scikit-train_test_split/
#sudo pip install imbalanced-learn
#https://machinelearningmastery.com/random-oversampling-and-undersampling-for-imbalanced-classification/

# example of random oversampling to balance the class distribution
from collections import Counter
from sklearn.datasets import make_classification
from imblearn.over_sampling import RandomOverSampler
from sklearn.model_selection import train_test_split
import pandas as pd
import numpy as np
from imblearn.over_sampling import RandomOverSampler, SMOTE


# Train dataset

mainfolder = "../data_2019_processed/"
outputFolders = "train/"

#import data
dataset = pd.read_csv(mainfolder+outputFolders+"allData-classification-numeric-normalized.csv")
#del dataset['showActuation']
#del dataset['distanceHome']

# Separate the class from the inputs
#X = dataset.iloc[:,:-1]
#y = dataset.iloc[:,-1]
y = dataset['class']
X = dataset.copy()
del X['class']
del X['timestamp']
del X['timestamp_text']

print("Train dataset")
# summarize class distribution
print("Before oversampling: ", Counter(y))

# define oversampling strategy
#oversample = RandomOverSampler(sampling_strategy='minority')
#oversample = RandomOverSampler(sampling_strategy='auto')
oversample = SMOTE(random_state = 32)

#y

#X


# fit and apply the transform
X_over, y_over = oversample.fit_resample(X, y)

# summarize class distribution
print("After undersampling: ", Counter(y_over))

dfBalanced = X_over
dfBalanced['class'] = y_over

dfBalanced

dfBalanced.to_csv(mainfolder+outputFolders+"allData-classification-numeric-normalized_balanced_oversample_smote.csv", sep=',', encoding='utf-8', index=False)

# Test dataset

outputFolders = "test/"

#import data
dataset = pd.read_csv(mainfolder+outputFolders+"allData-classification-numeric-normalized.csv")
#del dataset['showActuation']
#del dataset['distanceHome']

# Separate the class from the inputs
#X = dataset.iloc[:,:-1]
#y = dataset.iloc[:,-1]
y = dataset['class']
X = dataset.copy()
del X['class']
del X['timestamp']
del X['timestamp_text']

print("Train dataset")
# summarize class distribution
print("Before oversampling: ", Counter(y))
# define oversampling strategy
#oversample = RandomOverSampler(sampling_strategy='auto') #minority
oversample = SMOTE(random_state = 32)
# fit and apply the transform
X_over, y_over = oversample.fit_resample(X, y)
# summarize class distribution
print("After undersampling: ", Counter(y_over))

dfBalanced = X_over
dfBalanced['class'] = y_over

dfBalanced

dfBalanced.to_csv(mainfolder+outputFolders+"allData-classification-numeric-normalized_balanced_oversample_smote.csv", sep=',', encoding='utf-8', index=False)

# general data


outputFolders = ""

#import data
dataset = pd.read_csv(mainfolder+outputFolders+"allData-classification-numeric-normalized.csv")
#del dataset['showActuation']
#del dataset['distanceHome']

# Separate the class from the inputs
#X = dataset.iloc[:,:-1]
#y = dataset.iloc[:,-1]
y = dataset['class']
X = dataset.copy()
del X['class']
del X['timestamp']
del X['timestamp_text']

print("General data dataset")
# summarize class distribution
print("Before oversampling: ", Counter(y))
# define oversampling strategy
#oversample = RandomOverSampler(sampling_strategy='auto') #minority
oversample = SMOTE(random_state = 32)
# fit and apply the transform
X_over, y_over = oversample.fit_resample(X, y)
# summarize class distribution
print("After undersampling: ", Counter(y_over))

dfBalanced = X_over
dfBalanced['class'] = y_over

dfBalanced

dfBalanced.to_csv(mainfolder+outputFolders+"allData-classification-numeric-normalized_balanced_oversample_smote.csv", sep=',', encoding='utf-8', index=False)
