==> 00_scripts_raw_data/00_data_preprocessing/06_DataCorrelation.ipynb <==
import csv
import json
import pandas as pd
from datetime import datetime
import numpy as np

mainfolder = "../data_2019_processed/"
outputFolders = ""

# All data correlation

dataframe = pd.read_csv(mainfolder+outputFolders+"allData_filtred.csv", delimiter=',', encoding="utf8")
#dataframe = pd.read_csv("dados2019/student_a9Qgj8ENWrHvl9QqlXcIPKmyGMKgbfHk9Dbqon1HQP4.csv", delimiter=',', encoding="utf8")

# Exibe info
print(dataframe.info())

# convert data and show the new info
# activity
dataframe['activity'] = dataframe['activity'].replace(['Unknown'], '1')
dataframe['activity'] = dataframe['activity'].replace(['Still'], '2')
dataframe['activity'] = dataframe['activity'].replace(['Tilting'], '3')
dataframe['activity'] = dataframe['activity'].replace(['Exercise'], '4')
dataframe['activity'] = dataframe['activity'].replace(['In vehicle'], '5')
# class
dataframe['class'] = dataframe['class'].replace(['awake'], '0')
dataframe['class'] = dataframe['class'].replace(['asleep'], '1')
# day_of_week
dataframe['day_of_week'] = dataframe['day_of_week'].replace(['Tuesday'], '1')
dataframe['day_of_week'] = dataframe['day_of_week'].replace(['Wednesday'], '2')
dataframe['day_of_week'] = dataframe['day_of_week'].replace(['Thursday'], '3')
dataframe['day_of_week'] = dataframe['day_of_week'].replace(['Friday'], '4')
dataframe['day_of_week'] = dataframe['day_of_week'].replace(['Saturday'], '5')
dataframe['day_of_week'] = dataframe['day_of_week'].replace(['Sunday'], '6')
dataframe['day_of_week'] = dataframe['day_of_week'].replace(['Monday'], '7')
# phone_lock
dataframe['phone_lock'] = dataframe['phone_lock'].replace(['not locked'], '0')
dataframe['phone_lock'] = dataframe['phone_lock'].replace(['locked'], '1')
# location
dataframe['location'] = dataframe['location'].replace(['Other'], '1')
dataframe['location'] = dataframe['location'].replace(['University'], '2')
dataframe['location'] = dataframe['location'].replace(['House'], '3')

# timestamp
if 'timestamp2' in dataframe.columns:
    dataframe['timestamp_text'] = dataframe['timestamp']
    dataframe['timestamp'] = dataframe['timestamp2']
    del dataframe['timestamp2']
#dataframe['timestamp'] = dataframe['timestamp'].astype('datetime64[ns]')
#dataframe['timestamp'] = dataframe['timestamp'].astype('int64')
#dataframe['timestamp'] = datetime.strptime(dataframe['timestamp'],'%Y-%m-%d %H:%M:%S').timestamp() * 1000
#dataframe['timestamp'] = pd.to_datetime(dataframe['timestamp'],'%Y-%m-%d %H:%M:%S') * 1000
#dataframe['timestamp'] = dataframe['timestamp'].dt.second()
# activity
dataframe['activity'] = dataframe['activity'].astype(np.integer) 
dataframe['class'] = dataframe['class'].astype(np.integer) 
dataframe['day_of_week'] = dataframe['day_of_week'].astype(np.integer) 
dataframe['phone_lock'] = dataframe['phone_lock'].astype(np.integer) 
dataframe['location'] = dataframe['location'].astype(np.integer) 
# Exibe info
print(dataframe.info())# convert values

dataframe

 #'pearson', "   "'spearman', 'kendall'
a = dataframe['class'].corr(dataframe['activity'])
b = dataframe['class'].corr(dataframe['location'])
c = dataframe['class'].corr(dataframe['distanceHome'])
d = dataframe['class'].corr(dataframe['showActuation'])
e = dataframe['class'].corr(dataframe['timestamp'])
f = dataframe['class'].corr(dataframe['day_of_week'])
g = dataframe['class'].corr(dataframe['light'])
h = dataframe['class'].corr(dataframe['phone_lock'])
i = dataframe['class'].corr(dataframe['proximity'])
j = dataframe['class'].corr(dataframe['sound'])
l = dataframe['class'].corr(dataframe['time_to_next_alarm'])
m = dataframe['class'].corr(dataframe['sleep_time'])
n = dataframe['class'].corr(dataframe['wakeup_time'])
p = dataframe['class'].corr(dataframe['minutes_day'])
o = dataframe['class'].corr(dataframe['class'])
print("Correlations")
corrrrrrr = (("activity",a),
            ("location",b),
            ("distanceHome",c),
            ("showActuation",d),
            ("timestamp",e),
            ("day_of_week",f),
            ("light",g),
            ("phone_lock",h),
            ("proximity",i),
            ("sound",j),
            ("sleep_time",m),
            ("wakeup_time",n),
            ("time_to_next_alarm",l),
            ("minutes_day",p),
            ("class",o))
print (corrrrrrr)

type(corrrrrrr)

tests = pd.DataFrame(corrrrrrr,columns=("Feature","Correlation"))
tests.sort_values(by=['Correlation'])

# Only training correlation

outputFolders = "train/"

dataframe = pd.read_csv(mainfolder+outputFolders+"allData.csv", delimiter=',', encoding="utf8")
#dataframe = pd.read_csv("dados2019/student_a9Qgj8ENWrHvl9QqlXcIPKmyGMKgbfHk9Dbqon1HQP4.csv", delimiter=',', encoding="utf8")

# convert data and show the new info
# activity
dataframe['activity'] = dataframe['activity'].replace(['Unknown'], '1')
dataframe['activity'] = dataframe['activity'].replace(['Still'], '2')
dataframe['activity'] = dataframe['activity'].replace(['Tilting'], '3')
dataframe['activity'] = dataframe['activity'].replace(['Exercise'], '4')
dataframe['activity'] = dataframe['activity'].replace(['In vehicle'], '5')
# class
dataframe['class'] = dataframe['class'].replace(['awake'], '0')
dataframe['class'] = dataframe['class'].replace(['asleep'], '1')
# day_of_week
dataframe['day_of_week'] = dataframe['day_of_week'].replace(['Tuesday'], '1')
dataframe['day_of_week'] = dataframe['day_of_week'].replace(['Wednesday'], '2')
dataframe['day_of_week'] = dataframe['day_of_week'].replace(['Thursday'], '3')
dataframe['day_of_week'] = dataframe['day_of_week'].replace(['Friday'], '4')
dataframe['day_of_week'] = dataframe['day_of_week'].replace(['Saturday'], '5')
dataframe['day_of_week'] = dataframe['day_of_week'].replace(['Sunday'], '6')
dataframe['day_of_week'] = dataframe['day_of_week'].replace(['Monday'], '7')
# phone_lock
dataframe['phone_lock'] = dataframe['phone_lock'].replace(['not locked'], '0')
dataframe['phone_lock'] = dataframe['phone_lock'].replace(['locked'], '1')
# location
dataframe['location'] = dataframe['location'].replace(['Other'], '1')
dataframe['location'] = dataframe['location'].replace(['University'], '2')
dataframe['location'] = dataframe['location'].replace(['House'], '3')

# timestamp
if 'timestamp2' in dataframe.columns:
    dataframe['timestamp_text'] = dataframe['timestamp']
    dataframe['timestamp'] = dataframe['timestamp2']
    del dataframe['timestamp2']
#dataframe['timestamp'] = datetime.strptime(dataframe['timestamp'],'%Y-%m-%d %H:%M:%S').timestamp() * 1000
#dataframe['timestamp'] = pd.to_datetime(dataframe['timestamp'],'%Y-%m-%d %H:%M:%S') * 1000
#dataframe['timestamp'] = dataframe['timestamp'].dt.second()
# activity
dataframe['activity'] = dataframe['activity'].astype(np.integer) 
dataframe['class'] = dataframe['class'].astype(np.integer) 
dataframe['day_of_week'] = dataframe['day_of_week'].astype(np.integer) 
dataframe['phone_lock'] = dataframe['phone_lock'].astype(np.integer) 
dataframe['location'] = dataframe['location'].astype(np.integer) 
# Exibe info
print(dataframe.info())# convert values

 #'pearson', "   "'spearman', 'kendall'
a = dataframe['class'].corr(dataframe['activity'])
b = dataframe['class'].corr(dataframe['location'])
c = dataframe['class'].corr(dataframe['distanceHome'])
d = dataframe['class'].corr(dataframe['showActuation'])
e = dataframe['class'].corr(dataframe['timestamp'])
f = dataframe['class'].corr(dataframe['day_of_week'])
g = dataframe['class'].corr(dataframe['light'])
h = dataframe['class'].corr(dataframe['phone_lock'])
i = dataframe['class'].corr(dataframe['proximity'])
j = dataframe['class'].corr(dataframe['sound'])
l = dataframe['class'].corr(dataframe['time_to_next_alarm'])
m = dataframe['class'].corr(dataframe['sleep_time'])
n = dataframe['class'].corr(dataframe['wakeup_time'])
p = dataframe['class'].corr(dataframe['minutes_day'])
o = dataframe['class'].corr(dataframe['class'])
print("Correlations")
corrrrrrr = (("activity",a),
            ("location",b),
            ("distanceHome",c),
            ("showActuation",d),
            ("timestamp",e),
            ("day_of_week",f),
            ("light",g),
            ("phone_lock",h),
            ("proximity",i),
            ("sound",j),
            ("sleep_time",m),
            ("wakeup_time",n),
            ("time_to_next_alarm",l),
            ("minutes_day",p),
            ("class",o))
print (corrrrrrr)

tests = pd.DataFrame(corrrrrrr,columns=("Feature","Correlation"))
tests.sort_values(by=['Correlation'])
